backend/requirements.txt
# Core dependencies
Flask==2.3.2
Flask-CORS==4.0.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0

# Video processing
opencv-python==4.8.0.74
numpy==1.24.3
matplotlib==3.7.2
ffmpeg-python==0.2.0

# AI & ML
anthropic==0.28.0

# Google Cloud Platform
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.1
google-cloud-firestore==2.11.1
google-cloud-secret-manager==2.16.4

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.1.0
mypy==1.5.0

# Utilities
Pillow==10.0.0
